Add informer in place of controller and split code
